[controller][test] Split VersionLifecyclePolicyTest along its source-class sections
The consolidated test class was 750 lines / 43 tests covering seven helpers
across four conceptually distinct sections of VersionLifecyclePolicy.
Splitting along those same section boundaries gives four cohesive files
that will travel intact to the right home when the planned per-helper
extraction (TopicLifecycleManager, VersionRetirementManager,
VersionProvisioner, CurrentVersionManager) lands:

  NewPushCapacityGuardsTest    19 tests — backup-cleanup + rollback-origin gu
ards
  VersionSelectionTest          4 tests — getBackupVersionNumber
  PushStatusAggregationTest    11 tests — getOverallPushStatus + ge
tFinalReturnStatus
  LifecyclePredicatesTest       9 tests — hasFatalDataValidationError + up
dateStoreTTLRepushFlag

Shared mock factories (mockVersion, mockStoreWithVersions,
mockStoreWithPromoteTimestamp, setOf) and constants (CLUSTER_NAME,
STORE_NAME, MIN_*, ROLLED_BACK_RETENTION_MS) move to a package-private
VersionLifecycleTestSupport so the four files reuse them via explicit
static imports (no wildcards).

Test bodies are unchanged except for one trailing reset(store, repository)
removed from unrelatedPushIdDoesNotAffectTTLFlag — it was a no-op at the
end of an already-independent @Test method, preserved purely for parity
with the deleted in-admin host test. Total test count holds at 43; all
pass; spotbugs reports zero findings on the five new classes.
