+++ services/venice-controller/src/test/java/com/linkedin/venice/controller/versionlifecycle/NewPushCapacityGuardsTest.java
@@ -1,50 +1,39 @@
 package com.linkedin.venice.controller.versionlifecycle;
 
-import static com.linkedin.venice.meta.Store.NON_EXISTING_VERSION;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyBoolean;
+import static com.linkedin.venice.controller.versionlifecycle.VersionLifecycleTestSupport.CLUSTER_NAME;
+import static com.linkedin.venice.controller.versionlifecycle.VersionLifecycleTestSupport.MIN_CLEANUP_DELAY_MS;
+import static com.linkedin.venice.controller.versionlifecycle.VersionLifecycleTestSupport.MIN_VERSIONS_TO_PRESERVE;
+import static com.linkedin.venice.controller.versionlifecycle.VersionLifecycleTestSupport.ROLLED_BACK_RETENTION_MS;
+import static com.linkedin.venice.controller.versionlifecycle.VersionLifecycleTestSupport.STORE_NAME;
+import static com.linkedin.venice.controller.versionlifecycle.VersionLifecycleTestSupport.mockStoreWithPromoteTimestamp;
+import static com.linkedin.venice.controller.versionlifecycle.VersionLifecycleTestSupport.mockStoreWithVersions;
+import static com.linkedin.venice.controller.versionlifecycle.VersionLifecycleTestSupport.mockVersion;
 import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertThrows;
 import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.expectThrows;
 
 import com.linkedin.venice.exceptions.VeniceException;
 import com.linkedin.venice.meta.BackupStrategy;
-import com.linkedin.venice.meta.ReadWriteStoreRepository;
 import com.linkedin.venice.meta.Store;
 import com.linkedin.venice.meta.Version;
 import com.linkedin.venice.meta.VersionStatus;
-import com.linkedin.venice.pushmonitor.ExecutionStatus;
-import com.linkedin.venice.pushmonitor.OfflinePushStatus;
-import com.linkedin.venice.pushmonitor.PushMonitor;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import org.testng.annotations.Test;
 
 
-public class VersionLifecyclePolicyTest {
-  private static final String CLUSTER_NAME = "test-cluster";
-  private static final String STORE_NAME = "test-store";
-  private static final int MIN_VERSIONS_TO_PRESERVE = 2;
-  private static final long MIN_CLEANUP_DELAY_MS = TimeUnit.HOURS.toMillis(1);
-  private static final long ROLLED_BACK_RETENTION_MS = TimeUnit.HOURS.toMillis(24);
-
+/**
+ * Tests for the two new-push capacity guards on {@link VersionLifecyclePolicy}:
+ * {@code checkBackupVersionCleanupCapacityForNewPush} (backup-cleanup min-delay) and
+ * {@code checkRollbackOriginVersionCapacityForNewPush} (rollback-origin retention window).
+ */
+public class NewPushCapacityGuardsTest {
   // ---------- checkBackupVersionCleanupCapacityForNewPush ----------
-
   @Test
   public void deleteOnNewPushStartUsesNMinusOnePreserveCountAndPassesWhenNoVersionsPending() {
     // DELETE_ON_NEW_PUSH_START: after SOP cleanup we expect only current + new push.
@@ -422,332 +411,5 @@ public class VersionLifecyclePolicyTest {
         VeniceException.class,
         () -> VersionLifecyclePolicy
             .checkRollbackOriginVersionCapacityForNewPush(CLUSTER_NAME, STORE_NAME, store, retention, now));
-  }
-
-  // ---------- getBackupVersionNumber ----------
-
-  @Test
-  public void getBackupVersionNumberReturnsLargestOnlineBelowCurrent() {
-    List<Version> versions = Arrays.asList(
-        mockVersion(1, VersionStatus.ONLINE),
-        mockVersion(2, VersionStatus.ONLINE),
-        mockVersion(3, VersionStatus.ONLINE)); // current
-
-    assertEquals(VersionLifecyclePolicy.getBackupVersionNumber(versions, 3), 2);
-  }
-
-  @Test
-  public void getBackupVersionNumberSkipsNonOnlineVersions() {
-    // v3 is current; v2 is ERROR — should be skipped — v1 ONLINE wins
-    List<Version> versions = Arrays.asList(
-        mockVersion(1, VersionStatus.ONLINE),
-        mockVersion(2, VersionStatus.ERROR),
-        mockVersion(3, VersionStatus.ONLINE));
-
-    assertEquals(VersionLifecyclePolicy.getBackupVersionNumber(versions, 3), 1);
-  }
-
-  @Test
-  public void getBackupVersionNumberReturnsNonExistingWhenNoOnlineBelowCurrent() {
-    List<Version> versions = Arrays.asList(mockVersion(1, VersionStatus.ERROR), mockVersion(2, VersionStatus.ONLINE));
-
-    assertEquals(VersionLifecyclePolicy.getBackupVersionNumber(versions, 2), NON_EXISTING_VERSION);
-  }
-
-  @Test
-  public void getBackupVersionNumberWalksFromHighestToLowest() {
-    // Sort happens in-place desc; the first ONLINE below current wins.
-    List<Version> versions = Arrays.asList(
-        mockVersion(5, VersionStatus.ONLINE), // current
-        mockVersion(4, VersionStatus.ONLINE), // should win
-        mockVersion(3, VersionStatus.ONLINE));
-
-    assertEquals(VersionLifecyclePolicy.getBackupVersionNumber(versions, 5), 4);
-  }
-
-  // ---------- hasFatalDataValidationError ----------
-
-  @Test
-  public void hasFatalDataValidationErrorReturnsTrueWhenPushReportsError() {
-    PushMonitor pushMonitor = mock(PushMonitor.class);
-    OfflinePushStatus offlinePushStatus = mock(OfflinePushStatus.class);
-    doReturn(true).when(offlinePushStatus).hasFatalDataValidationError();
-    doReturn(offlinePushStatus).when(pushMonitor).getOfflinePushOrThrow("topic_v1");
-
-    assertTrue(VersionLifecyclePolicy.hasFatalDataValidationError(pushMonitor, "topic_v1"));
-  }
-
-  @Test
-  public void hasFatalDataValidationErrorReturnsFalseWhenPushReportsClean() {
-    PushMonitor pushMonitor = mock(PushMonitor.class);
-    OfflinePushStatus offlinePushStatus = mock(OfflinePushStatus.class);
-    doReturn(false).when(offlinePushStatus).hasFatalDataValidationError();
-    doReturn(offlinePushStatus).when(pushMonitor).getOfflinePushOrThrow("topic_v1");
-
-    assertFalse(VersionLifecyclePolicy.hasFatalDataValidationError(pushMonitor, "topic_v1"));
-  }
-
-  @Test
-  public void hasFatalDataValidationErrorReturnsFalseWhenPushEntryMissing() {
-    PushMonitor pushMonitor = mock(PushMonitor.class);
-    doThrow(new VeniceException("not found")).when(pushMonitor).getOfflinePushOrThrow("topic_v1");
-
-    assertFalse(VersionLifecyclePolicy.hasFatalDataValidationError(pushMonitor, "topic_v1"));
-  }
-
-  // ---------- updateStoreTTLRepushFlag ----------
-
-  @Test
-  public void ttlRePushSetsFlagToTrueWhenCurrentlyFalse() {
-    Store store = mock(Store.class);
-    ReadWriteStoreRepository repository = mock(ReadWriteStoreRepository.class);
-    String ttlRepushId = Version.generateTTLRePushId("test-push");
-    when(store.isTTLRepushEnabled()).thenReturn(false);
-
-    VersionLifecyclePolicy.updateStoreTTLRepushFlag(ttlRepushId, store, repository);
-
-    verify(store).setTTLRepushEnabled(true);
-    verify(repository).updateStore(store);
-  }
-
-  @Test
-  public void ttlRePushIsNoOpWhenFlagAlreadyTrue() {
-    Store store = mock(Store.class);
-    ReadWriteStoreRepository repository = mock(ReadWriteStoreRepository.class);
-    String ttlRepushId = Version.generateTTLRePushId("test-push");
-    when(store.isTTLRepushEnabled()).thenReturn(true);
-
-    VersionLifecyclePolicy.updateStoreTTLRepushFlag(ttlRepushId, store, repository);
-
-    verify(store, never()).setTTLRepushEnabled(anyBoolean());
-    verify(repository, never()).updateStore(any());
-  }
-
-  @Test
-  public void regularPushWithTTLClearsFlagWhenCurrentlyTrue() {
-    Store store = mock(Store.class);
-    ReadWriteStoreRepository repository = mock(ReadWriteStoreRepository.class);
-    String regularPushWithTtlId = Version.generateRegularPushWithTTLRePushId("test-push");
-    when(store.isTTLRepushEnabled()).thenReturn(true);
-
-    VersionLifecyclePolicy.updateStoreTTLRepushFlag(regularPushWithTtlId, store, repository);
-
-    verify(store).setTTLRepushEnabled(false);
-    verify(repository).updateStore(store);
-  }
-
-  @Test
-  public void regularPushWithTTLIsNoOpWhenFlagAlreadyFalse() {
-    Store store = mock(Store.class);
-    ReadWriteStoreRepository repository = mock(ReadWriteStoreRepository.class);
-    String regularPushWithTtlId = Version.generateRegularPushWithTTLRePushId("test-push");
-    when(store.isTTLRepushEnabled()).thenReturn(false);
-
-    VersionLifecyclePolicy.updateStoreTTLRepushFlag(regularPushWithTtlId, store, repository);
-
-    verify(store, never()).setTTLRepushEnabled(anyBoolean());
-    verify(repository, never()).updateStore(any());
-  }
-
-  @Test
-  public void compliancePushDoesNotAffectTTLFlag() {
-    Store store = mock(Store.class);
-    ReadWriteStoreRepository repository = mock(ReadWriteStoreRepository.class);
-    String compliancePushId = Version.generateCompliancePushId("test-push");
-    when(store.isTTLRepushEnabled()).thenReturn(true);
-
-    VersionLifecyclePolicy.updateStoreTTLRepushFlag(compliancePushId, store, repository);
-
-    verify(store, never()).setTTLRepushEnabled(anyBoolean());
-    verify(repository, never()).updateStore(any());
-  }
-
-  @Test
-  public void unrelatedPushIdDoesNotAffectTTLFlag() {
-    Store store = mock(Store.class);
-    ReadWriteStoreRepository repository = mock(ReadWriteStoreRepository.class);
-    String userPushId = System.currentTimeMillis() + "_https://example.com/user-push";
-    when(store.isTTLRepushEnabled()).thenReturn(true);
-
-    VersionLifecyclePolicy.updateStoreTTLRepushFlag(userPushId, store, repository);
-
-    verify(store, never()).setTTLRepushEnabled(anyBoolean());
-    verify(repository, never()).updateStore(any());
-    reset(store, repository); // keep parity with the original test file
-  }
-
-  // ---------- getOverallPushStatus ----------
-
-  @Test
-  public void overallPushStatusReturnsCompletedWhenBothCompleted() {
-    assertEquals(
-        VersionLifecyclePolicy.getOverallPushStatus(ExecutionStatus.COMPLETED, ExecutionStatus.COMPLETED),
-        ExecutionStatus.COMPLETED);
-  }
-
-  @Test
-  public void overallPushStatusReturnsErrorWhenEitherIsError() {
-    assertEquals(
-        VersionLifecyclePolicy.getOverallPushStatus(ExecutionStatus.ERROR, ExecutionStatus.COMPLETED),
-        ExecutionStatus.ERROR);
-    assertEquals(
-        VersionLifecyclePolicy.getOverallPushStatus(ExecutionStatus.ERROR, ExecutionStatus.ERROR),
-        ExecutionStatus.ERROR);
-  }
-
-  @Test
-  public void overallPushStatusPrefersDvcIngestionErrorOverError() {
-    // DVC ingestion errors sit higher in STATUS_PRIORITIES than ERROR, so they outrank a Venice ERROR.
-    assertEquals(
-        VersionLifecyclePolicy
-            .getOverallPushStatus(ExecutionStatus.COMPLETED, ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL),
-        ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-    assertEquals(
-        VersionLifecyclePolicy
-            .getOverallPushStatus(ExecutionStatus.ERROR, ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL),
-        ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-  }
-
-  // ---------- getFinalReturnStatus ----------
-
-  @Test
-  public void finalReturnStatusAllCompletedReturnsCompleted() {
-    Set<String> regions = setOf("region1", "region2", "region3");
-    Map<String, ExecutionStatus> statuses = new HashMap<>();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.COMPLETED);
-    statuses.put("region3", ExecutionStatus.COMPLETED);
-
-    assertEquals(
-        VersionLifecyclePolicy.getFinalReturnStatus(statuses, regions, 0, new StringBuilder()),
-        ExecutionStatus.COMPLETED);
-  }
-
-  @Test
-  public void finalReturnStatusProgressOutranksCompleted() {
-    Set<String> regions = setOf("region1", "region2", "region3");
-    Map<String, ExecutionStatus> statuses = new HashMap<>();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.PROGRESS);
-    statuses.put("region3", ExecutionStatus.COMPLETED);
-
-    assertEquals(
-        VersionLifecyclePolicy.getFinalReturnStatus(statuses, regions, 0, new StringBuilder()),
-        ExecutionStatus.PROGRESS);
-  }
-
-  @Test
-  public void finalReturnStatusErrorBeatsCompleted() {
-    Set<String> regions = setOf("region1", "region2", "region3");
-    Map<String, ExecutionStatus> statuses = new HashMap<>();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.ERROR);
-    statuses.put("region3", ExecutionStatus.COMPLETED);
-
-    assertEquals(
-        VersionLifecyclePolicy.getFinalReturnStatus(statuses, regions, 0, new StringBuilder()),
-        ExecutionStatus.ERROR);
-  }
-
-  @Test
-  public void finalReturnStatusUnknownBeatsErrorWithOneRegionFailed() {
-    Set<String> regions = setOf("region1", "region2", "region3");
-    Map<String, ExecutionStatus> statuses = new HashMap<>();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.ERROR);
-    statuses.put("region3", ExecutionStatus.UNKNOWN);
-
-    assertEquals(
-        VersionLifecyclePolicy.getFinalReturnStatus(statuses, regions, 1, new StringBuilder()),
-        ExecutionStatus.UNKNOWN);
-  }
-
-  @Test
-  public void finalReturnStatusFailedMajorityDowngradesToProgress() {
-    // 2 of 3 regions unreachable → strict majority of 2 not met → downgrade to PROGRESS.
-    Set<String> regions = setOf("region1", "region2", "region3");
-    Map<String, ExecutionStatus> statuses = new HashMap<>();
-    statuses.put("region1", ExecutionStatus.UNKNOWN);
-    statuses.put("region2", ExecutionStatus.ERROR);
-    statuses.put("region3", ExecutionStatus.UNKNOWN);
-
-    assertEquals(
-        VersionLifecyclePolicy.getFinalReturnStatus(statuses, regions, 2, new StringBuilder()),
-        ExecutionStatus.PROGRESS);
-  }
-
-  @Test
-  public void finalReturnStatusDvcDiskFullOutranksDvcOther() {
-    Set<String> regions = setOf("region1", "region2", "region3");
-    Map<String, ExecutionStatus> statuses = new HashMap<>();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.COMPLETED);
-    statuses.put("region3", ExecutionStatus.DVC_INGESTION_ERROR_OTHER);
-
-    assertEquals(
-        VersionLifecyclePolicy.getFinalReturnStatus(statuses, regions, 0, new StringBuilder()),
-        ExecutionStatus.DVC_INGESTION_ERROR_OTHER);
-
-    statuses.put("region3", ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-    assertEquals(
-        VersionLifecyclePolicy.getFinalReturnStatus(statuses, regions, 0, new StringBuilder()),
-        ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-  }
-
-  @Test
-  public void finalReturnStatusDvcDiskFullOutranksMemoryLimit() {
-    Set<String> regions = setOf("region1", "region2", "region3");
-    Map<String, ExecutionStatus> statuses = new HashMap<>();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.DVC_INGESTION_ERROR_MEMORY_LIMIT_REACHED);
-    statuses.put("region3", ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-
-    assertEquals(
-        VersionLifecyclePolicy.getFinalReturnStatus(statuses, regions, 0, new StringBuilder()),
-        ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-  }
-
-  @Test
-  public void finalReturnStatusDvcOtherOutranksMemoryLimit() {
-    Set<String> regions = setOf("region1", "region2", "region3");
-    Map<String, ExecutionStatus> statuses = new HashMap<>();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.DVC_INGESTION_ERROR_MEMORY_LIMIT_REACHED);
-    statuses.put("region3", ExecutionStatus.DVC_INGESTION_ERROR_OTHER);
-
-    assertEquals(
-        VersionLifecyclePolicy.getFinalReturnStatus(statuses, regions, 0, new StringBuilder()),
-        ExecutionStatus.DVC_INGESTION_ERROR_OTHER);
-  }
-
-  // ---------- helpers ----------
-
-  private static Version mockVersion(int number, VersionStatus status) {
-    Version v = mock(Version.class);
-    doReturn(number).when(v).getNumber();
-    doReturn(status).when(v).getStatus();
-    return v;
-  }
-
-  private static Store mockStoreWithVersions(int currentVersion, long promotedMsAgo, Version... versions) {
-    Store store = mock(Store.class);
-    doReturn(currentVersion).when(store).getCurrentVersion();
-    doReturn(System.currentTimeMillis() - promotedMsAgo).when(store).getLatestVersionPromoteToCurrentTimestamp();
-    doReturn(versions.length == 0 ? Collections.emptyList() : Arrays.asList(versions)).when(store).getVersions();
-    return store;
-  }
-
-  private static Store mockStoreWithPromoteTimestamp(int currentVersion, long promoteTimestampMs, Version... versions) {
-    Store store = mock(Store.class);
-    doReturn(currentVersion).when(store).getCurrentVersion();
-    doReturn(promoteTimestampMs).when(store).getLatestVersionPromoteToCurrentTimestamp();
-    doReturn(versions.length == 0 ? Collections.emptyList() : Arrays.asList(versions)).when(store).getVersions();
-    return store;
-  }
-
-  private static Set<String> setOf(String... values) {
-    Set<String> out = new HashSet<>();
-    Collections.addAll(out, values);
-    return out;
   }
 }
